Upgrades can now be found randomly by destroying soft walls.
The two possible upgrades are a blast and a carry boost. The first one
increases the blast radius of the player's bombs, whereas the second one
increases the total amount of live bombs the player may have at the same
time.

The upgrades vanish after a certain amount of time, or can be destroyed
by an explosion.