Add explosion sprites.
Explosions get assigned a certain sprite based on their direction. The
origin will branch out in all four directions. North and South sprites
get a vertical explosion sprite, whereas east and west get a horizontal
one.

The end of each explosion is determined by the detonate signal's
strength. If it is zero, we know that it's the last tile in to draw.