Added a very basic AI which will work as a foundation for the future.
First of all I removed the old player and foe classes. Instead I created
the Entity parent class from which both NPCs and players now inherit.
Basically NPCs and players are equal except for the methods they use to
handle their actions. While the player classes are driven by user input
the NPCs generate their own input. Since they all use the same commands
it should be easy to create different kinds of AI (agressive, passive,
etc.) later on.

I also changed the way that bombs mark tiles as dangerous for the AI.
They now add their own danger values to an existing value and remove it
again on explosion. The system is still a bit bugged, especially when a
bomb has been kicked, but it works good enough to build on it.

For testing purposes the game now has one player controlled character,
which starts in the top left corner and three AI controlled characters
which start in the other corners of the arena.

The AI currently walks around randomly, plants bombs when they are next
to softwalls and try to evade bombs.