NPCs now try to reach the closest upgrade they can find.
To achieve this I had to find some way to let the NPCs have access to
the currently active upgrades. Due to the way I handle the upgrades
(and other tile content) that was harder than I imagined. The ugly way
would have been to give the npcs more access to the whole game grid and
let them search the upgrades themselves. This would've been the brute
force approach.

I now settled for an upgrade manager where upgrades are registered and
removed on the fly. It also finds the closest upgrade from the list to
a given set of coordinates.