NPCs are controlled by a simple state manager.
The only two states at the moment are Walk and Evade. Walk handles the
main "thinking" so to speak as it checks for good bomb spots and the
decisions the NPC should make. I'll probably split it up into other
states in the future.

Evading does what the name implicates. It moves the NPC away from a
dangerous tile, by checking the adjacent tiles and finding a safe spot.
The algorithm I used here is self made, but somewhat inspired by the A*
pathfinding concept.

When the NPC is not running away from a dangerous tile it will hunt for
the player or for an upgrade (depending on which one of the two is
closer to the npc). This still is a very early draft of the feature and
uses pretty random movement instead of some more complex pathfinding.
Players are now listed in a PlayerManager which allows easy access and
makes it easy to find a player close to the npc.

The NPC will place bombs only when it is close to a softwall and does
NOT check for nearby players or wether it is safe to place the bomb or
not.