Improved explosion code by moving it outside of the tile class.
The sprite for an explosion tile is now picked by the explosion objects
themselves instead of by the parent tile. Apart from giving the code
a better structure it also made it easier to create a more elaborate
sprite picking algorithm.

Explosions work the same way as before, but I was able to fix the bug,
where the end tile of an explosion would interrupt another explosion,
creating ugly artifacts. The new code now checks if the adjacent tile
is also an explosion before deciding wether to place an end tile or not.
This is possible now, because the explosion's appearance is not set in
stone at the time it is create, but constantly re-evaluated over its
lifetime.

This also allows smooth transitions from intersecting to single
explosion tiles.

Last but not least I fixed the bug where old explosions weren't replaced
by new ones. The new system feels much more reasonable now.