Fix Issue #5 - Load images and set their filters separately.
The issue was caused, because the default filter for images was set
after the images were loaded.

All resources are now loaded at the beginning of the game with the help
of a resource manager. This means the löve's default filter will affect
them now.

The resource manager should also come in handy when I need to implement
music and sounds.