Added a preliminary main menu which uses the player to navigate the buttons.
The main menu is comprised of a small arena, a player, a few buttons and an ugly
logo. The player can select buttons by moving the character up and down.
Camera tracking is activated.

On top of that the game now starts on the main menu.

We have seperated input maps for the game and the main menu.

The shader for switching the game's colours has been moved from main.love to the
main menu and game screens, where it makes more sense.