All static tiles are now drawn by the arena.
Moved the code for drawing all the static objects (floor, hardwalls and
softwalls) to the arena class.

To make it easier to switch between different level art the sprites are
now be loaded into different tables named after the level they are made
for. Switching between these tables will switch the artwork for the
level.