New screens are initialized after they have been pushed onto the stack.
This fixes an issue, where popping a screen in its init() function would
derail the screen manager, since the pop() function would affect the
screen on top of the stack instead of the newly created one.

Now this is prevented, since the init() function is only called after
the new screen has already been put onto the stack.