Add destruction animation for soft walls.
Whenever a soft wall is destroyed and no upgrade is dropped, we place
a destruction tile instead. A destruction tile is nothing more than a
capsule for an animation which will be played.

The type of the animation depends on the tileset of the current level
and is picked when the destruction tile is created.

After the animation has been played once, the tile is removed.