Fix #7 and #13 - Fix movement issues by implementing AABB collision.
Improved the movement code by checking the bounding boxes of adjacent
tiles versus the entities bounding box. It now continues to move towards
an impassable tile until their bounding boxes collide.

Before this fix the entity would often stop too early, because the
collision check already reported a collision even when the entity wasn't
on the current tile completely. This often lead to entities getting
stuck in between tiles.

A side effect of the above fix is, that it also got rid of the bomb
kicking issue, where bombs would start moving too early. Now the
entity's bbox has to collide with the bombs bbox to initiate the bomb
kicking.