Fix #4, #6, #10 - General AI improvement.
First of all I got rid of the idle state. We now have states for movement,
random movement and evasion. The AI will only stop if it is really necessary.

The random movement state will find a random direction to move to as long as it
is passable and safe. When the npc comes to a crossroad it might pick a new
direction or continue along the old path. This state is timed. We generally
don't want the NPC to stay here too long, because it is too random. Currently
the random state will only be entered, when the NPC can't move closer to its
target (set by the move state) because its path is blocked.

The movement state will - like before - aquire a target to which the npc will
try to move, by always following the tiles with the smallest movement costs.
As mentioned above it will change to the random mode as soon as that path is
blocked.

As before, the evasion state is entered, as soon as the NPC steps onto a tile
which isn't safe. The evasion state will look for a nearby safe tile and move
the npc there. I added a few more safety measures and also excluded the previous
movement direction from the evasion path. This might lead to trapped NPCs but
all in all improved the movement.

Another big improvement concerns re #3. The random state will move the npc until it is
completely on the target tile, which makes the overall movement feel more natural. Re #3
is still not fixed though. The issue is still especially noticeable in the evasion state.