Add endboss for the first level.
The boss is an entity at heart but has some special abilities. For
example it doesn't plant bombs, but instead shoots missiles across the
arena and releases minions (which randomly walk around the grid). It can
walk over bombs, but stops them, when they are kicked towards it.

It has two states. The first one makes it walk horizontally across the
screen until it reaches a wall. The boss then randomly decides wether to
move two rows up or down. While walking, it can release minions and
missiles (timebased).

If it is hurt by an explosion it will enter the hurt state, which makes
it pulsate and invincible for a short amount of time.

When the boss is beat, the next level is loaded. Therefore some changes
had to be applied to the Arena and to the LevelSwitcher classes.