Rewrote most of the level code to make it more modular.
The biggest change is that now most of the actual level related stuff
is only reset (e.g. players are moved to their spawns, the arena is
cleared and refilled with softwalls, etc.) and not actually recreated.
This reduced the garbage produced between rounds a lot.

The actual changing between rounds and stages now happens in the
LevelSwitcher class. It crashes after the last round of stage three
has been finished, but this shouldn't be a problem since we will add
a boss battle as stage four anyway.