Entities aren't removed from their managers upon death.
Entities are now marked as dead instead of being removed from their
respective managers. Since we only use the getCount() function to see
if all npcs or players are dead and a round should be ended, it wasn't
really necessary to remove the entities.

Instead we now just exclude the dead ones from the getCount() function.

The good thing is, that we could reuse them for the next round, stage
or level and don't have to create new objects all the time.