Improve level selection menu to lock levels and use teleporters.
I changed the layout of the menu to make space for six levels in total.
Instead of selecting the levels like a button, the player now can walk
into the preview images to teleport to that level. Therefore I removed
all the button logic. Instead I added AABB collision to see if the
player has entered a certain portal.

Removing the buttons also made it easier / more convenient to add a
mechanism for locking and unlocking levels. Basically each teleporter is
"hidden" behind a door. By default only the door to the first level is
open. All the others will have to be unlocked by defeating bosses.

The information about each door's lock-state is saved to the player
profile. This way we can track the progress over multiple game sessions.
The added ProfileHandler should also come in handy when we want to store
other information in the future.

A small fix was applied to the wave shader. It now considers the alpha
channel of the texture it is applied to.