Make screens globally available to fix issues with requiring them.
Basically the problem was, that due to how the screen manager works,
screens couldn't easily switch back to another screen.

For example:
Main Menu -> Options -> Main Menu

This would create an issue, where Main Menu needs to require the Options
screen, but the Options menu also needs to require the Main Menu to
switch back to it. This would have created a require loop and therefore
crashed the game.

Of course the better fix would be to rewrite the screen manager, but
making the screens globally available is much more simple and also
doesn't really have any negative effects on the performance / game.