Add proper outro screen which displays the current scores.
The outro screen shows the scores of all rounds on the current
stage. Add an animated dodger face that smiles / frowns and win / lose
jingles based on the score of the last round.

The level screen is responsible for playing and stopping the music at
the right times.

Also removed the IntroScreen since it became superfluous.