Improve menus and general game layout.
The game is now limited to a resolution of 640x480 pixels. The
background now is filled with the default gameboy color. Camera tracking
now follows the center of an entity instead of the top left corner of
its sprite.

The camera movement is now also limited on the level menu. I also
removed the teleporters and doors for levels 5 and 6. The game will only
feature 4 levels.