Add activation / deactivation of shaders.
Shaders will be deactivated if the game runs on machines that don't
support shaders (the option for toggling shaders also won't work on
those machines).

Shaders use a small wrapper class now, which produces shader objects.
These wrappers take care of setting shaders when they are enabled and
ignoring them when they are disabled.

Also enabled the color switcher for both the option and the main menu
and changed the background rectangle to fit the gameboy palette (in case
shaders are deactivated).